Move logic from coverage to subgraph mod
